+++ stream-manipulation/webrtc-streaming/client/index.html
@@ -64,15 +64,12 @@
             <input id="cam_height" name="cam_height" type="number" value="300"/>
             <label for="nn_model">Neural Network</label>
             <select name="nn_model" id="nn_model">
-                <option value="">--</option>ä
-                <option value="face-detection-adas-0001">face-detection-adas-0001</option>
-                <option value="face-detection-retail-0004">face-detection-retail-0004</option>
-                <option value="mobilenet-ssd">mobilenet-ssd</option>
-                <option value="pedestrian-and-vehicle-detector-adas-0001">pedestrian-and-vehicle-detector-adas-0001</option>
-                <option value="pedestrian-detection-adas-0002">pedestrian-detection-adas-0002</option>
-                <option value="person-detection-retail-0013">person-detection-retail-0013</option>
-                <option value="person-vehicle-bike-detection-crossroad-1016">person-vehicle-bike-detection-crossroad-1016</option>
-                <option value="vehicle-detection-adas-0002">vehicle-detection-adas-0002</option>
+                <option value="">--</option>
+                <option value="luxonis/yolov6-nano:r2-coco-512x288">luxonis/yolov6-nano:r2-coco-512x288</option>
+                <option value="luxonis/yunet:640x480">luxonis/yunet:640x480</option>
+                <option value="luxonis/ppe-detection:640x640">luxonis/ppe-detection:640x640</option>
+                <option value="luxonis/mediapipe-palm-detection:192x192">luxonis/mediapipe-palm-detection:192x192</option>
+                <option value="luxonis/scrfd-person-detection:25g-640x640">luxonis/scrfd-person-detection:25g-640x640</option>
             </select>
         </div>
         <div id="depth_options" style="display: none;">
@@ -82,17 +79,15 @@
                 <option value="THE_720_P">THE_720_P</option>
                 <option value="THE_800_P">THE_800_P</option>
             </select>
-            <label for="median_filter">Median Filter</label>
-            <select name="median_filter" id="median_filter">
-                <option value="MEDIAN_OFF">MEDIAN_OFF</option>
-                <option value="KERNEL_3x3">KERNEL_3x3</option>
-                <option value="KERNEL_5x5">KERNEL_5x5</option>
-                <option value="KERNEL_7x7" selected>KERNEL_7x7</option>
+            <label for="preset_mode">Stereo Depth Preset Mode</label>
+            <select name="preset_mode" id="preset_mode">
+                <option value="HIGH_ACCURACY">HIGH_ACCURACY</option>
+                <option value="DEFAULT">DEFAULT</option>
+                <option value="HIGH_DENSITY">HIGH_DENSITY</option>
+                <option value="FACE">FACE</option>
+                <option value="HIGH_DETAIL">HIGH_DETAIL</option>
+                <option value="ROBOTICS">ROBOTICS</option>
             </select>
-            <input type="checkbox" name="subpixel" value="on">
-            <label for="subpixel">Subpixel</label>
-            <input type="checkbox" name="extended_disparity" value="on">
-            <label for="extended_disparity">Extended disparity</label>
         </div>
     </form>
 </div>
