+++ stream-manipulation/webrtc-streaming/utils/options_wrapper.py
@@ -1,0 +1,27 @@
+class OptionsWrapper:
+    def __init__(self, raw_options):
+        self.raw_options = raw_options
+
+    @property
+    def camera_type(self):
+        return self.raw_options.get("camera_type", "rgb")
+
+    @property
+    def width(self):
+        return int(self.raw_options.get("cam_width", 300))
+
+    @property
+    def height(self):
+        return int(self.raw_options.get("cam_height", 300))
+
+    @property
+    def nn(self):
+        return self.raw_options.get("nn_model", "")
+
+    @property
+    def mono_camera_resolution(self):
+        return self.raw_options.get("mono_camera_resolution", "THE_400_P")
+
+    @property
+    def preset_mode(self):
+        return self.raw_options.get("preset_mode", "HIGH_ACCURACY")
